refactor Cmi5 class
